home work

+++ api/src/Controller/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Category;
 use App\Entity\Product;
+use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Config\Definition\Exception\Exception;
@@ -26,10 +27,20 @@ class ProductController extends AbstractController
         $this->entityManager = $entityManager;
     }
 
+    /**
+     * @param Request $request
+     * @return JsonResponse
+     * @throw Exception
+     */
     #[Route('/product-create', name: 'product_create')]
     public function create(Request $request): JsonResponse
     {
+        /** @var $requestData */
         $requestData = json_decode($request->getContent(), true);
+
+        /** @var User $user */
+        $user = $this->getUser();
+
         if (!isset(
             $requestData['price'],
             $requestData['name'],
@@ -39,11 +50,14 @@ class ProductController extends AbstractController
             throw new Exception('Invalid request data');
         }
 
+        /** @var $category */
         $category = $this->entityManager->getRepository(Category::class)->find($requestData["category"]);
+
         if (!$category) {
             throw new Exception("Category with id " . $requestData['category'] . " not found");
         }
 
+        /** @var $product */
         $product = new Product();
 
         $product
@@ -52,24 +66,32 @@ class ProductController extends AbstractController
             ->setDescription($requestData['description'])
             ->setCategory($category);
 
-        $this->entityManager->persist($product);
-        $this->entityManager->flush();
+        if (in_array(User::ROLE_ADMIN, $user->getRoles())) {
+            $this->entityManager->persist($product);
+            $this->entityManager->flush();
 
-        return new JsonResponse($product, Response::HTTP_CREATED);
+            return new JsonResponse($product, Response::HTTP_CREATED);
+        } else {
+            return new JsonResponse("You are not Admin", Response::HTTP_FORBIDDEN);
+        }
     }
 
+    /**
+     * @return JsonResponse
+     */
     #[Route('/products', name: 'product_get_all')]
     public function getAll(): JsonResponse
     {
         /** @var Product $product */
         $products = $this->entityManager->getRepository(Product::class)->findAll();
 
-        return new JsonResponse($products);
+        return new JsonResponse($products, Response::HTTP_OK);
     }
 
     /**
      * @param string $id
      * @return JsonResponse
+     * @throw Exception
      */
     #[Route('product/{id}', name: 'product_get_item')]
     public function getItem(string $id): JsonResponse
@@ -81,29 +103,64 @@ class ProductController extends AbstractController
             throw new Exception("Product with id: " . $id . " not found");
         }
 
-        return new JsonResponse($product);
+        return new JsonResponse($product, Response::HTTP_OK);
     }
 
+    /**
+     * @param Request $request
+     * @param string $id
+     * @return JsonResponse
+     */
     #[Route('product-update/{id}', name: 'product_update_item')]
-    public function updateProduct(string $id): JsonResponse
+    public function updateProduct(Request $request ,string $id): JsonResponse
     {
         /** @var Product $product */
         $product = $this->entityManager->getRepository(Product::class)->find($id);
+
+        /** @var User $user */
+        $user = $this->getUser();
 
         if (!$product) {
             throw new Exception("Product with id: " . $id . " not found");
         }
 
-        $product->setName("new name");
+        /** @var $requestData */
+        $requestData = json_decode($request->getContent(), true);
 
-        $this->entityManager->flush();
+        /** @var $fieldsToUpdate */
+        $fieldsToUpdate = ['name', 'price', 'description'];
 
-        return new JsonResponse($product);
+        foreach ($fieldsToUpdate as $field) {
+            if (isset($requestData[$field])) {
+                $setterMethod = 'set' . ucfirst($field);
+                $product->$setterMethod($requestData[$field]);
+            }
+        }
+
+        $category = $this->entityManager->getRepository(Category::class)->find($requestData["category"]);
+
+        $product
+            ->setCategory($category);
+
+        if (in_array(User::ROLE_ADMIN, $user->getRoles())) {
+            $this->entityManager->flush();
+
+            return new JsonResponse($product, Response::HTTP_OK);
+        } else {
+            return new JsonResponse("You are not Admin", Response::HTTP_FORBIDDEN);
+        }
     }
 
+    /**
+     * @param string $id
+     * @return JsonResponse
+     */
     #[Route('product-delete/{id}', name: 'product_delete_item')]
     public function deleteProduct(string $id): JsonResponse
     {
+        /** @var User $user */
+        $user = $this->getUser();
+
         /** @var Product $product */
         $product = $this->entityManager->getRepository(Product::class)->find($id);
 
@@ -111,11 +168,38 @@ class ProductController extends AbstractController
             throw new Exception("Product with id: " . $id . " not found");
         }
 
-        $product->setName("new name");
+        if (in_array(User::ROLE_ADMIN, $user->getRoles())) {
+            $this->entityManager->remove($product);
+            $this->entityManager->flush();
 
-        $this->entityManager->remove($product);
-        $this->entityManager->flush();
+            return new JsonResponse('', Response::HTTP_NO_CONTENT);
+        } else {
+            return new JsonResponse("You are not Admin", Response::HTTP_FORBIDDEN);
+        }
+    }
 
-        return new JsonResponse();
+    /**
+     * @param Request $request
+     * @return JsonResponse
+     */
+    #[Route(path: "product-find", name: "app_find_product")]
+    public function findProduct(Request $request): JsonResponse
+    {
+        $requestData = $request->query->all();
+
+        /** @var Product $product */
+        $product = $this->entityManager->getRepository(Category::class)->getAllProducts(
+            $requestData['itemsPerPage'] ?? 10,
+            $requestData['page'] ?? 1,
+            $requestData['id'] ?? null,
+            $requestData['categoryId'] ?? null,
+            $requestData['categoryId'] ?? null,
+            $requestData['categoryType'] ?? null,
+            $requestData['name'] ?? null,
+            $requestData['price'] ?? null,
+            $requestData['description'] ?? null
+        );
+
+        return new JsonResponse($product, Response::HTTP_OK);
     }
 }
